refactor(workflow-runtime): 매칭 JSON fallback 관측성을 보장
