+++ .agent/specs/608.md
@@ -1,0 +1,74 @@
+# Workflow Runtime 매칭 책임 분리
+
+## Goal
+
+`WorkflowMatchingService`가 담당하던 매칭 오케스트레이션, 후보 조회, 점수 산정, decision 기록, metric, JSON fallback 책임을 분리해 기존 매칭 결과는 유지하면서 route/intent JSON parse 실패를 관측 가능하게 만든다.
+
+## Problem
+
+현재 `WorkflowMatchingService`는 세션 조회, redaction, signal guard, query embedding, 후보 검색, rerank, decision 기록, metric 처리를 한 클래스에서 수행한다. 또한 route/intent 조건 JSON이나 workflow meta JSON을 파싱할 때 실패하면 빈 객체로 조용히 대체되어, 매칭 품질 저하가 JSON 입력 문제인지 점수 산정 문제인지 추적하기 어렵다.
+
+## Scope
+
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingService.java`는 매칭 유스케이스 오케스트레이션 중심으로 유지한다.
+- matching core, profile candidate provider, decision recorder, metric 책임을 `backend/src/main/java/com/init/workflowruntime/application/matching/` 내부 collaborator로 분리한다.
+- route/intent 조건 JSON 및 matching profile text 생성에 쓰이는 JSON parse fallback 정책을 공통화한다.
+- JSON parse 실패 시 raw JSON을 노출하지 않고 로그와 metric으로 fallback 발생 source를 관측할 수 있게 한다.
+- 기존 matching result status, selected workflow/intent, decision log 기록 의미는 보존한다.
+- `backend/src/test/java/com/init/workflowruntime/application/matching/`의 단위 테스트로 기존 동작과 malformed JSON fallback 관측을 확인한다.
+
+## Non-goals
+
+- 매칭 점수 가중치, threshold 기본값, 후보 조회 SQL, embedding provider/model을 변경하지 않는다.
+- `runtime.workflow_match_decision` 테이블 스키마나 저장 payload 계약을 변경하지 않는다.
+- workflow runtime presentation API, WebSocket, workflow execution 동작을 변경하지 않는다.
+- frontend 또는 ML pipeline 코드는 변경하지 않는다.
+
+## Affected Paths
+
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingService.java`
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingProfileTextFactory.java`
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingCore.java`
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingEvaluation.java`
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingJsonParser.java`
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingMetrics.java`
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingProfileCandidateProvider.java`
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingTextSignals.java`
+- `backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchDecisionRecorder.java`
+- `backend/src/test/java/com/init/workflowruntime/application/matching/WorkflowMatchingServiceTest.java`
+- `backend/src/test/java/com/init/workflowruntime/application/matching/WorkflowMatchingProfileTextFactoryTest.java`
+- `backend/src/test/java/com/init/workflowruntime/application/matching/WorkflowMatchingProfileBuildWorkerTest.java`
+
+## Requirements
+
+1. `WorkflowMatchingService.match(...)`는 기존처럼 embedding 비활성화, insufficient context, profile missing, successful matching, runtime error fallback을 처리해야 한다.
+2. 후보 count, query embedding cache, nearest active search, lexical search query 생성은 profile candidate provider 책임으로 분리되어야 한다.
+3. rerank, route/intent entry condition 평가, confusion type 결정, final result decision은 matching core 책임으로 분리되어야 한다.
+4. decision log JSON 직렬화와 failure reason 결정은 decision recorder 책임으로 분리되어야 한다.
+5. matching metric 증가는 metric 전용 collaborator로 분리되어야 한다.
+6. route/intent JSON parse 실패는 빈 객체 fallback을 유지하되, fallback source를 로그와 `workflow_matching.json_fallback` metric tag로 남겨야 한다.
+7. malformed JSON fallback 때문에 전체 match가 ERROR로 끝나면 안 되며, 기존 scoring fallback 의미와 decision log 기록은 유지되어야 한다.
+
+## Data/API Impact
+
+- REST API, WebSocket API, DB schema 변경은 없다.
+- `runtime.workflow_match_decision` 기록 필드와 기존 JSON payload 구조는 유지한다.
+- 새로운 metric `workflow_matching.json_fallback{source=...}`가 추가된다.
+
+## Acceptance Criteria
+
+- `WorkflowMatchingService` 생성자와 본문에서 profile repository, decision repository, `ObjectMapper`, `MeterRegistry`, `Clock` 세부 책임이 직접 섞이지 않는다.
+- route/intent JSON parse 실패 시 match는 fallback으로 계속 진행하고, JSON fallback metric이 증가한다.
+- `WorkflowMatchingProfileTextFactory`도 동일 parser를 사용해 malformed JSON을 조용히 삼키지 않는다.
+- 기존 `WorkflowMatchingServiceTest`의 주요 matching 시나리오가 계속 통과한다.
+- JSON parse 실패 케이스를 확인하는 단위 테스트가 추가된다.
+
+## Validation
+
+- `cd backend && ./gradlew test --tests com.init.workflowruntime.application.matching.WorkflowMatchingServiceTest --tests com.init.workflowruntime.application.matching.WorkflowMatchingProfileTextFactoryTest`
+- `cd backend && ./gradlew test --tests com.init.workflowruntime.application.matching.*`
+- `git diff --check`
+
+## Open Questions
+
+- 없음. 이슈 본문과 참고 위치만으로 backend application-layer refactor 범위와 fallback 관측 요구가 결정된다.
