+++ backend/src/main/java/com/init/workflowruntime/application/matching/WorkflowMatchingEvaluation.java
@@ -1,0 +1,6 @@
+package com.init.workflowruntime.application.matching;
+
+import java.util.List;
+
+public record WorkflowMatchingEvaluation(
+    WorkflowMatchResult result, List<WorkflowMatchCandidate> rankedCandidates) {}
