+++ backend/src/test/java/com/init/workflowruntime/application/matching/WorkflowMatchingProfileTextFactoryTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.init.domainpack.domain.model.IntentDefinition;
 import com.init.domainpack.domain.model.WorkflowDefinition;
+import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +13,8 @@ import org.junit.jupiter.api.Test;
 class WorkflowMatchingProfileTextFactoryTest {
 
   private final WorkflowMatchingProfileTextFactory factory =
-      new WorkflowMatchingProfileTextFactory(new ObjectMapper());
+      new WorkflowMatchingProfileTextFactory(
+          new WorkflowMatchingJsonParser(new ObjectMapper(), new SimpleMeterRegistry()));
 
   @Test
   @DisplayName("raw JSON dump 대신 route/evidence/quality를 구조화한 profile text를 만든다")
@@ -55,5 +57,41 @@ class WorkflowMatchingProfileTextFactoryTest {
     assertThat(text).contains("workflow_steps: start, 주문 확인");
     assertThat(text).contains("workflow_replay_fitness: 0.82");
     assertThat(text).contains("lexical_terms:");
+  }
+
+  @Test
+  @DisplayName("profile text 생성 중 malformed JSON은 빈 객체 fallback과 metric으로 관측한다")
+  void should_recordJsonFallbackMetric_when_profileJsonIsMalformed() {
+    SimpleMeterRegistry meterRegistry = new SimpleMeterRegistry();
+    WorkflowMatchingProfileTextFactory textFactory =
+        new WorkflowMatchingProfileTextFactory(
+            new WorkflowMatchingJsonParser(new ObjectMapper(), meterRegistry));
+    IntentDefinition intent =
+        IntentDefinition.create(
+            101L, "refund_request", "환불 요청", "고객이 결제 환불을 요청한다.", 1, "{}", "{bad-json", "[]", "{}");
+    WorkflowDefinition workflow =
+        WorkflowDefinition.create(
+            101L,
+            "refund_flow",
+            "환불 접수",
+            "환불 가능 여부를 확인하고 접수한다.",
+            "{}",
+            "start",
+            "[]",
+            "[]",
+            "{}",
+            10L,
+            true,
+            "{}");
+
+    String text = textFactory.build(intent, workflow);
+
+    assertThat(text).contains("intent_code: refund_request");
+    assertThat(
+            meterRegistry
+                .counter(
+                    "workflow_matching.json_fallback", "source", "profile_intent_entry_condition")
+                .count())
+        .isEqualTo(1.0);
   }
 }
